Extensive changes (and some bug fixes) for the CSV related functionality

+++ trunk/src/org/yooreeka/examples/higgs/HiggsBoson.java
@@ -67,7 +67,7 @@ public class HiggsBoson {
 	 * @return 
 	 * 
 	 */
-	public CSVEntry query(long id) {
+	public CSVEntry query(int id) {
 		return loader.getF().getDoc().getCsvData().get(id);	
 	}
